# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vertex AI Mailer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/google/callback:
    get:
      operationId: googleAuthCallback
      tags: [auth]
      summary: Google OAuth callback
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect after auth

  # ─── GMAIL ──────────────────────────────────────────────────────────────────
...
  /ai/generate-email:
    post:
      operationId: generateEmail
      tags: [ai]
      summary: AI-generate a personalized email for a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateEmailInput"
      responses:
        "200":
          description: Generated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedEmail"
  /ai/preview-email:
    post:
      operationId: previewEmail
      tags: [ai]
      summary: Preview how a template renders for a specific lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreviewEmailInput"
      responses:
        "200":
          description: Rendered email preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedEmail"

  # ─── DRAFTS ─────────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Admin — list all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
...
  /admin/logs:
    get:
      operationId: adminGetLogs
      tags: [admin]
      summary: Admin — draft generation and AI usage logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: System logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLogs"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        gmailConnected:
          type: boolean
        gmailEmail:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        aiTone:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, email, name, role, gmailConnected, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [email, password]
    RegisterInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
          minLength: 1
      required: [email, password, name]

    # ─── GMAIL ───────────────────────────────────────────────────────────────
...
    GenerateEmailInput:
      type: object
      properties:
        leadId:
          type: integer
        templateId:
          type: integer
        tone:
          type: string
          enum: [professional, friendly, sales, followup, urgent]
          default: professional
        customPrompt:
          type: ["string", "null"]
      required: [leadId, templateId]
    PreviewEmailInput:
      type: object
      properties:
        templateId:
          type: integer
        leadData:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
            vehicle:
              type: ["string", "null"]
            route:
              type: ["string", "null"]
            pickup:
              type: ["string", "null"]
            delivery:
              type: ["string", "null"]
            price:
              type: ["string", "null"]
          required: [name, email]
        tone:
          type: string
          enum: [professional, friendly, sales, followup, urgent]
          default: professional
      required: [templateId, leadData]
    GeneratedEmail:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        tone:
          type: string
      required: [subject, body, tone]

    # ─── DRAFTS ──────────────────────────────────────────────────────────────
...
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    SystemLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [draft_created, draft_failed, ai_call, gmail_error, upload_error]
        description:
          type: string
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
      required: [id, type, description, createdAt]
    PaginatedLogs:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SystemLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]